backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.3.5
anthropic>=0.20.0
pypdf==3.17.1
python-pptx==0.6.23
python-docx==1.1.0
